[ofono-ril-plugin] Treat no call as a remote hangup.
On an incoming call, when the caller hangup
before pickup the call on device, the reason
is set as OFONO_DISCONNECT_REASON_ERROR, which
makes the device emitting three tones at full
blow.

This is due to the following exchange in the
modem:
- slot1 > 2 callStateChanged
- slot1 < [00000157] 10 getCurrentCalls
- slot1 > [00000157] 140 getCurrentCallsResponse_1_2
- slot1 < [0000015b] 19 getLastCallFailCause
- slot1 > [0000015b] 18 getLastCallFailCauseResponse

The current calls response contains no call,
so self->calls in ril_voicecall.c is emptied.
Then, the getLastCallFailCauseResponse, gives
CALL_FAIL_NORMAL_UNSPECIFIED as a reason,
so the code looks for the status of the call. Which
has been deleted from the list.